requirements.txt
PySimpleGUI
matplotlib
Pillow